+++ PHEngine/PHEngine/src/Core/GameCore/Event/PhysicsCollisionEvent.h
@@ -18,10 +18,10 @@ namespace Event
 
    using ParentEvent_t = TEvent<eEventThreadType::GAME_THREAD, MultipleDataEventPolicy<ePhysicsCollisionStateType,
                                                         ePhysicsBodyType,
-                                                        uint32_t /*this_phys_descriptor_id*/,
-                                                        uint64_t /*this_actor_id*/,
-                                                        uint32_t /*that_phys_descriptor_id*/,
-                                                        uint64_t /*that_actor_id*/>>;
+                                                        int32_t /*this_phys_descriptor_id*/,
+                                                        int32_t /*this_actor_id*/,
+                                                        int32_t /*that_phys_descriptor_id*/,
+                                                        int32_t /*that_actor_id*/>>;
 
    class PhysicsCollisionEvent
        : public ParentEvent_t
