+++ PHEngine/PHEngine/src/Core/GraphicsCore/SceneProxy/SceneProxyBase.h
@@ -1,18 +1,19 @@
 #pragma once
 
 #include <cstddef>
+#include <stdint.h>
 
 namespace Graphics {
 
    class SceneProxyBase
    {
-      static size_t SceneProxyIdCounter;
+      static int32_t SceneProxyIdCounter;
 
    protected:
 
-      size_t mSceneProxyId;
+      int32_t mSceneProxyId;
 
-      size_t mGameObjectId;
+      int32_t mGameObjectId;
 
       bool mIsEnabled;
 
@@ -22,15 +23,15 @@ namespace Graphics {
 
       virtual ~SceneProxyBase();
 
-      size_t GetSceneProxyId() const;
+      int32_t GetSceneProxyId() const;
 
       void SetEnabled(const bool bEnabled);
 
       bool IsEnabled() const;
 
-      void SetBindedGameObjectId(const size_t gameObjectId);
+      void SetBindedGameObjectId(const int32_t gameObjectId);
 
-      size_t GetGameObjectId() const;
+      int32_t GetGameObjectId() const;
 
    };
 }
