+++ PHEngine/PHEngine/src/Core/Job.h
@@ -10,17 +10,17 @@ namespace Thread
    {
       using callback_t = std::function<void(void)>;
 
-      uint64_t mCreatorObjectId;
+      int32_t mCreatorObjectId;
       uint64_t mFunctionId;
       callback_t mCallback;
 
    public:
 
-      Job(const uint64_t creatorObjectId, const uint64_t functionId, callback_t callback);
+      Job(const int32_t creatorObjectId, const uint64_t functionId, callback_t callback);
 
       ~Job();
 
-      uint64_t GetCreatorObjectId() const;
+      int32_t GetCreatorObjectId() const;
       uint64_t GetFunctionId() const;
       callback_t GetCallback() const;
